refactor(controllers): extract McpActivityTracker from HelpSessionController
- Move MCP tool-call activity strip, tier-mismatch approval flow, per-tool
  grant lifecycle, session-revoked banner, and turn-outcome alert pip into
  a sibling module — including its own IPC listener start()/dispose()
- Controller delegates through a narrow host interface (getSnapshot/patch)
  — no behavior change
